requirements.txt
beautifulsoup4~=4.12.3
hnswlib==0.8.0
loguru==0.7.2
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==3.4.1
torch==2.6.0
transformers==4.40.2

click~=8.1.7
numpy~=1.26.4
FlagEmbedding~=1.3.4
scikit-learn~=1.3.2
tqdm~=4.66.6
evaluate~=0.4.3
matplotlib~=3.10.0
pytest~=8.3.4